refactor!: hide RecordMappingStrategy from the public API
The class is a vendored OpenCSV helper used by a single caller, ReportDataEndpoint, and never meant for library users. Moving it out of the csv package and making it package-private keeps it out of the published javadoc and out of the API surface.

BREAKING CHANGE: dev.andstuff.kraken.api.endpoint.account.csv.RecordMappingStrategy is no longer public, which makes the next release a breaking one for anyone who referenced it.
