+++ library/src/main/java/dev/andstuff/kraken/api/endpoint/account/RecordMappingStrategy.java
@@ -1,4 +1,4 @@
-package dev.andstuff.kraken.api.endpoint.account.csv;
+package dev.andstuff.kraken.api.endpoint.account;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -23,26 +23,12 @@ import com.opencsv.exceptions.CsvRuntimeException;
  *
  * @param <T> the record type each CSV line is mapped to
  */
-public class RecordMappingStrategy<T extends Record> extends HeaderColumnNameMappingStrategy<T> {
+class RecordMappingStrategy<T extends Record> extends HeaderColumnNameMappingStrategy<T> {
 
-    /**
-     * Creates a strategy mapping CSV lines to the given record type.
-     *
-     * @param type the record type each CSV line is mapped to
-     */
-    public RecordMappingStrategy(Class<T> type) {
+    RecordMappingStrategy(Class<T> type) {
         setType(type);
     }
 
-    /**
-     * Builds a record from a CSV line, calling its canonical constructor.
-     *
-     * @param line the values of the CSV line
-     * @return the record built from the line
-     * @throws CsvBeanIntrospectionException if a value cannot be read
-     * @throws CsvFieldAssignmentException if a value cannot be converted
-     * @throws CsvRuntimeException if the line doesn't have as many values as the record has components, or if the record cannot be instantiated
-     */
     @Override
     public T populateNewBean(String[] line) throws CsvBeanIntrospectionException, CsvFieldAssignmentException {
         RecordComponent[] recordComponents = type.getRecordComponents();
@@ -89,12 +75,6 @@ public class RecordMappingStrategy<T extends Record> extends HeaderColumnNameMap
      */
     public static class KrakenInstantConverter extends AbstractCsvConverter {
 
-        /**
-         * Converts a Kraken date and time to an {@link Instant}.
-         *
-         * @param value the value read from the CSV file
-         * @return the corresponding instant
-         */
         @Override
         public Object convertToRead(String value) {
             String[] dateTime = value.split(" ");
